create neuron_to_arbor module responsible of conversion
